move trait ConvertPipelineIdToWebRender to a better place
