+++ components/compositing/compositor.rs
@@ -28,8 +28,8 @@ use layers::platform::surface::NativeDisplay;
 use layers::rendergl;
 use layers::rendergl::RenderContext;
 use layers::scene::Scene;
-use layout_traits::LayoutControlChan;
-use msg::constellation_msg::{ConvertPipelineIdFromWebRender, ConvertPipelineIdToWebRender, Image, PixelFormat};
+use layout_traits::{ConvertPipelineIdToWebRender, LayoutControlChan};
+use msg::constellation_msg::{ConvertPipelineIdFromWebRender, Image, PixelFormat};
 use msg::constellation_msg::{Key, KeyModifiers, KeyState, LoadData};
 use msg::constellation_msg::{NavigationDirection, PipelineId, WindowSizeData, WindowSizeType};
 use pipeline::CompositionPipeline;
