+++ components/msg/constellation_msg.rs
@@ -365,20 +365,8 @@ impl fmt::Display for PipelineId {
 #[derive(Clone, PartialEq, Eq, Copy, Hash, Debug, Deserialize, Serialize, HeapSizeOf)]
 pub struct SubpageId(pub u32);
 
-pub trait ConvertPipelineIdToWebRender {
-    fn to_webrender(&self) -> webrender_traits::PipelineId;
-}
-
 pub trait ConvertPipelineIdFromWebRender {
     fn from_webrender(&self) -> PipelineId;
-}
-
-impl ConvertPipelineIdToWebRender for PipelineId {
-    fn to_webrender(&self) -> webrender_traits::PipelineId {
-        let PipelineNamespaceId(namespace_id) = self.namespace_id;
-        let PipelineIndex(index) = self.index;
-        webrender_traits::PipelineId(namespace_id, index)
-    }
 }
 
 impl ConvertPipelineIdFromWebRender for webrender_traits::PipelineId {
